fix(todo-item): corrected interaction with local-storage service

+++ src/todo-item/service/todo-item.service.ts
@@ -18,7 +18,7 @@ class TodoItemService implements ITodoItemService {
 
   async addTodo(dto: ITodoItemCreateFormDto): Promise<void> {
     const todo = TodoItemFactory.fromFormCreateDto(dto);
-    this._todos = [todo, ...this._todos];
+    this._todos = [...this._todos, todo];
 
     await this._repo.save(todo).catch(console.error);
   }
@@ -26,9 +26,9 @@ class TodoItemService implements ITodoItemService {
   async deleteTodo(todoId: string): Promise<void> {
     const idx = this._todos.findIndex((todo) => todo.id === todoId);
 
-    if (idx) {
-      this._todos.splice(idx, 1);
+    if (isFinite(idx) && idx !== -1) {
       const todo = this._todos[idx];
+      this._todos.splice(idx, 1);
       await this._repo.remove(todo).catch(console.error);
     }
   }
@@ -36,29 +36,36 @@ class TodoItemService implements ITodoItemService {
   async editTodo(todoId: string, dto: ITodoItemCreateFormDto): Promise<void> {
     const idx = this.todos.findIndex((item) => item.id === todoId);
 
-    if (idx) {
+    if (isFinite(idx) && idx !== -1) {
       const todo = this._todos[idx];
       const newTodo = TodoItemFactory.fromFormEditDto(dto, todo);
       this._todos.splice(idx, 1, newTodo);
+      await this._repo.remove(todo).catch(console.error);
       await this._repo.save(newTodo).catch(console.error);
     }
   }
 
   async toggleTodo(todoId: string): Promise<void> {
-    const idx = this.todos.findIndex((item) => item.id === todoId);
+    const idx = this._todos.findIndex((item) => item.id === todoId);
 
-    if (idx) {
+    if (isFinite(idx) && idx !== -1) {
       const todo = this._todos[idx];
       todo.completed = !todo.completed;
       this._todos.splice(idx, 1, todo);
+      await this._repo.remove(todo).catch(console.error);
       await this._repo.save(todo).catch(console.error);
     }
   }
 
   async init(todos: TodoItemModel[]): Promise<void> {
-    this._todos = todos;
+    const storageTodos = await this._repo.list().catch(console.error);
 
-    await Promise.all(this._todos.map((item) => this._repo.save(item))).catch(console.error);
+    if (!storageTodos?.length) {
+      this._todos = todos;
+      await Promise.all(this._todos.map((item) => this._repo.save(item))).catch(console.error);
+    } else {
+      this._todos = storageTodos;
+    }
   }
 }
